+++ ApartmentHouseManagement/AHM.WebAPI/Controllers/PackageTypeController.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using System.Web.Http;
 using AHM.BusinessLayer.Interfaces;
+using AHM.Common;
 using AHM.Common.DomainModel;
 using AHM.WebAPI.Attributes;
 
@@ -34,7 +35,7 @@ namespace AHM.WebAPI.Controllers
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(ModelState.SelectMany(m => m.Value.Errors).First().ErrorMessage);
             }
 
             if (AppUser.BuildingId.HasValue)
@@ -53,7 +54,7 @@ namespace AHM.WebAPI.Controllers
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(ModelState.SelectMany(m => m.Value.Errors).First().ErrorMessage);
             }
 
             var result = await _packageTypeService.UpdateAsync(packageType);
@@ -67,7 +68,13 @@ namespace AHM.WebAPI.Controllers
         {
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(ModelState.SelectMany(m => m.Value.Errors).First().ErrorMessage);
+            }
+
+            var inUse = await _packageTypeService.InUseAsync(packageType.Id);
+            if (inUse)
+            {
+                return BadRequest(ValidationMessages.PackageTypeInUse);
             }
 
             var result = await _packageTypeService.RemoveAsync(packageType.Id);
